backend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
    "tensorflow>=2.14.0",
    "textblob>=0.19.0",
    "trafilatura>=2.0.0",
    "vadersentiment>=3.3.2",
    "xgboost>=3.0.2",
    "flask>=3.0.3",
    "flask-cors>=5.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.9.0",
]
